Check if video exists before deleting
verinice-veo#3472

+++ cypress.config.js
@@ -31,7 +31,7 @@ module.exports = defineConfig({
         if (results && results.video) {
           // Do we have failures for any retry attempts?
           const failures = results.tests.some((test) => test.attempts.some((attempt) => attempt.state === 'failed'));
-          if (!failures) {
+          if (!failures && fs.existsSync(results.video)) {
             // delete the video if the spec passed and no tests retried
             fs.unlinkSync(results.video);
           }
